# subproyectoIntegracion-l8-2/src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: 'https://city-events-api.ew.r.appspot.com/api'
  - url: 'http://localhost:8089/api'
info:
  description: Api about cities and events that take place in them
  version: 1.0.0
  title: CityEventsAPI
  contact:
    email: javnunrui@alum.us.es
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /events:
    get:
      tags:
        - events
      summary: Get events
      description: Get all events with the right attributes
      operationId: getEvents
      parameters:
        - name: limit
          in: query
          description: The maximum number of items to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: name
          in: query
          description: Filter events by name
          required: false
          schema:
            type: string
        - name: organizer
          in: query
          description: Filter events by organizer
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter events by category
          required: false
          schema:
            type: string
        - name: location
          in: query
          description: Filter events by location
          required: false
          schema:
            type: string
        - name: isEmpty
          in: query
          description: Filter events with no fqas
          required: false
          schema:
            type: boolean
        - name: fields
          in: query
          description: Returns only the properties of the events
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: Order results by category, organizer or date
          required: false
          schema:
            type: string
            enum:
              - category
              - '-category'
              - organizer
              - '-organizer'
              - date
              - '-date'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    post:
      tags:
        - events
      summary: Create a event
      description: Create a event with the right attributes
      operationId: createEvent
      requestBody:
        description: Event to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    put:
      tags:
        - events
      summary: Update a event
      description: Update a event introducing new values for certain attributes
      operationId: updateEvent
      requestBody:
        description: Event to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '204':
          description: No content. Event updated
        '400':
          description: Bad request
        '404':
          description: Event not found
  '/events/{id}':
    get:
      tags:
        - events
      summary: Get a event by ID
      description: Get a event by ID with the right attributes
      operationId: getEvent
      parameters:
        - name: id
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event with corresponding id found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    delete:
      tags:
        - events
      summary: Delete a event
      description: Delete a event
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content. Event deleted
        '404':
          description: Event not found
  '/events/timeLeft/{id}':
    get:
      tags:
        - events
      summary: Get the remaining time for an event by ID
      description: Get the remaining time in days for an event by ID
      operationId: getTimeLeftEvent
      parameters:
        - name: id
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calculated remaining days for event with corresponding id
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Event not found
  '/events/{eventId}/{fqaId}':
    post:
      tags:
        - events
      summary: Add fqa to event
      description: Add fqa with fqaId to event with eventId
      operationId: addFqaToEvent
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters:
        - name: eventId
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
        - name: fqaId
          in: path
          description: Id of the fqa
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Fqa added to event correctly
          content:
            application/json:
              schema:
                required:
                  - name
                  - description
                  - organizer
                  - category
                  - location
                  - date
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  organizer:
                    type: string
                  category:
                    type: string
                  location:
                    type: string
                  date:
                    type: string
                  price:
                    type: string
                  fqas:
                    type: array
                    items:
                      type: object
                      required:
                        - question
                        - answer
                example:
                  id: e5
                  name: Iron Maiden Concert
                  description: Iron Maiden returns to Spain this summer
                  organizer: Delegacion de arte de Catalunya
                  category: Music
                  location: Estadi Olimpic Lluis Companys
                  date: '29/06/2022'
                  price: '100-200 euros'
                  fqas:
                    - id: f15
                      question: Area for reduced mobility?
                      answer: There is a conveniently equipped area for people with reduced mobility...
        '400':
          description: Bad request
        '404':
          description: Event or fqa not found
    delete:
      tags:
        - events
      summary: Delete fqa from event
      description: Delete fqa from event
      operationId: deleteFqaFromEvent
      parameters:
        - name: eventId
          in: path
          description: Id of the event
          required: true
          schema:
            type: string
        - name: fqaId
          in: path
          description: Id of the fqa
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content. Fqa deleted from event
        '404':
          description: Event or fqa not found
  /fqas:
    get:
      tags:
        - fqas
      summary: Get fqas
      description: Get all fqas with the right attributes
      operationId: getFqas
      parameters:
        - name: limit
          in: query
          description: The maximum number of items to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: fields
          in: query
          description: Returns only the properties of the fqas
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: Order results by question or answer
          required: false
          schema:
            type: string
            enum:
              - question
              - '-question'
              - answer
              - '-answer'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fqa'
        '400':
          description: Bad request
    post:
      tags:
        - fqas
      summary: Create a fqa
      description: Create a fqa with the rights attributes
      operationId: createFqa
      requestBody:
        description: Fqa to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fqa'
      responses:
        '201':
          description: Fqa created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fqa'
        '400':
          description: Bad request
    put:
      tags:
        - fqas
      summary: Update a fqa
      description: Update a fqa introducing new values for certain attributes
      operationId: updateFqa
      requestBody:
        description: Fqa to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fqa'
      responses:
        '204':
          description: No content. Fqa updated
        '404':
          description: Fqa not found
...
components:
  schemas:
...
    Event:
      type: object
      required:
        - name
        - description
        - organizer
        - category
        - location
        - date
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        organizer:
          type: string
        category:
          type: string
        location:
          type: string
        date:
          type: string
        price:
          type: string
        fqas:
          type: array
          items:
            $ref: '#/components/schemas/Fqa'
      example:
        id: e5
        name: Iron Maiden Concert
        description: Iron Maiden returns to Spain this summer
        organizer: Delegacion de arte de Catalunya
        category: Music
        location: Estadi Olimpic Lluis Companys
        date: '29/06/2022'
        price: '100-200 euros'
        fqas:
          - id: f15
            question: Area for reduced mobility?
            answer: There is a conveniently equipped area for people with reduced mobility...
    Fqa:
      type: object
      required:
        - question
        - answer
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
      example:
        id: f15
        question: Area for reduced mobility?
        answer: There is a conveniently equipped area for people with reduced mobility...